requirements.txt
antlr4-python3-runtime==4.13.1
black==24.3.0
importlib_resources==6.4.0
jsonschema==4.21.1
referencing==0.34.0
ruamel.yaml==0.18.6
